Create semi_automated_hashcat.sh
This will assume rockyou.txt is in the same folder you are running the script from. It will ask for you to paste a hash, try to determine the hash type, present you with hashcat's guess on the type, and then ask you to enter the "mode", which it will use to specify the expected hash type. It will then ask for the number of characters in the password if you know. Just hit enter if you don't know this.