Update monitor_d3d11.sh
Simple script to check for new arcdps from source website, hash it, compare that hash to included hash, and download new arcdps if the hashes don't match, and if they do, create a folder named by the hash and date/time and download the files into there, while retaining a copy elsewhere for future version comparisons. Just for fun.